% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d3circle.R
\name{d3circle}
\alias{d3circle}
\title{Zoomable Circle Packing}
\usage{
d3circle(data, settings = NULL, diameter = 500)
}
\arguments{
\item{diameter}{}
}
\description{
Zoomable Circle Packing
}
\examples{
require(d3circle)
require(jsonlite)

data <- fromJSON(system.file("data/flare.json",package='d3circle'),
                 simplifyDataFrame = FALSE)

d3circle(data = data,diameter=900)
}

